Replace the previous conc-build script
